Change ``` code fences to ~~~ for cross-compatibility
``` seems to be github specific so the docs dev server itself does not
pick them up properly. ~~~ mitigates the issue.

The docs server doesn't do anything particular with the "html" syntax
designator, but we may want to apply syntax highlighting later so
specifying it from the get-go seems efficient; and if viewing the md
files in github, this will ensure the syntax highlighting is correct.